# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateEvidenceEnvelope": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/EvidenceCandidate"
          },
          "counterexamples": {
            "$ref": "#/components/schemas/EvidenceCounterexamples"
          },
          "evidence_level": {
            "const": "candidate",
            "title": "Evidence Level"
          },
          "independence": {
            "$ref": "#/components/schemas/EvidenceIndependence"
          },
          "ledger": {
            "$ref": "#/components/schemas/EvidenceLedger"
          },
          "result": {
            "$ref": "#/components/schemas/EvidenceResult"
          },
          "schema_version": {
            "const": "evidence-envelope-v1",
            "title": "Schema Version"
          },
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          }
        },
        "required": [
          "schema_version",
          "evidence_level",
          "candidate",
          "source",
          "result",
          "independence",
          "counterexamples",
          "ledger"
        ],
        "title": "CandidateEvidenceEnvelope",
        "type": "object"
      },
...
      "CandidateReference": {
        "additionalProperties": false,
        "properties": {
          "candidate_note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Note"
          },
          "description": {
            "maxLength": 600,
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "maxLength": 120,
            "title": "Domain",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/CandidateEvidenceEnvelope"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retrieval_rank": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Retrieval Rank",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "domain",
          "description",
          "retrieval_rank",
          "evidence"
        ],
        "title": "CandidateReference",
        "type": "object"
      },
...
      "EvidenceCandidate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "score": {
            "title": "Score",
            "type": "null"
          },
          "status": {
            "const": "recorded",
            "title": "Status"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceCandidate",
        "type": "object"
      },
      "EvidenceCounterexamples": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": [
              "not_recorded",
              "gap_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "status"
        ],
        "title": "EvidenceCounterexamples",
        "type": "object"
      },
      "EvidenceIndependence": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "not_recorded",
            "title": "Kind"
          },
          "status": {
            "const": "not_recorded",
            "title": "Status"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceIndependence",
        "type": "object"
      },
      "EvidenceLedger": {
        "additionalProperties": false,
        "properties": {
          "artifact_sha256": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Sha256"
          },
          "claim_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Id"
          },
          "recorded_at": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "status": {
            "const": "not_recorded",
            "title": "Status"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "version": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "status"
        ],
        "title": "EvidenceLedger",
        "type": "object"
      },
      "EvidenceResult": {
        "additionalProperties": false,
        "properties": {
          "provenance": {
            "enum": [
              "NOT_TESTED",
              "INTERNAL_AI_SCREEN"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "status": {
            "enum": [
              "recorded",
              "not_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "verdict": {
            "enum": [
              "NOT_TESTED",
              "INCONCLUSIVE"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "status",
          "provenance",
          "verdict"
        ],
        "title": "EvidenceResult",
        "type": "object"
      },
      "EvidenceSource": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "not_recorded",
              "internal_kb"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "source_review": {
            "title": "Source Review",
            "type": "null"
          },
          "status": {
            "enum": [
              "recorded",
              "not_recorded"
            ],
            "title": "Status",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "null"
          }
        },
        "required": [
          "status",
          "kind"
        ],
        "title": "EvidenceSource",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LintClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^lint-[0-9a-f]{16}$",
            "title": "Claim Id",
            "type": "string"
          },
          "claim_type": {
            "enum": [
              "assumption",
              "analogy",
              "causal_judgment"
            ],
            "title": "Claim Type",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/CandidateEvidenceEnvelope"
          },
          "failure_mode": {
            "maxLength": 800,
            "minLength": 1,
            "title": "Failure Mode",
            "type": "string"
          },
          "quote": {
            "maxLength": 600,
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          },
          "reference_candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_priority": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Review Priority",
            "type": "string"
          },
          "structure": {
            "maxLength": 800,
            "minLength": 1,
            "title": "Structure",
            "type": "string"
          },
          "suggestion": {
            "maxLength": 800,
            "minLength": 1,
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "quote",
          "claim_type",
          "structure",
          "failure_mode",
          "review_priority",
          "suggestion",
          "reference_candidate",
          "evidence"
        ],
        "title": "LintClaim",
        "type": "object"
      },
...
      "SearchRequest": {
        "properties": {
          "lang": {
            "default": "zh",
            "description": "Output language for LLM-generated text: 'zh' or 'en'",
            "title": "Lang",
            "type": "string"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "rewrite": {
            "default": false,
            "description": "Use LLM to rewrite query for better matching",
            "title": "Rewrite",
            "type": "boolean"
          },
          "top_k": {
            "default": 12,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
...
      "StructLintRequest": {
        "additionalProperties": false,
        "description": "Body for POST /api/struct-lint.\n\n`document` carries the full plan text. The pydantic max_length is a\nstructural floor; check_doc_length() gives a friendlier JSON error\nwith the limit + received counts before we ever reach the LLM.",
        "properties": {
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 12,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{11,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "document": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Document",
            "type": "string"
          }
        },
        "required": [
          "document"
        ],
        "title": "StructLintRequest",
        "type": "object"
      },
      "StructLintResponse": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/LintClaim"
            },
            "maxItems": 30,
            "title": "Claims",
            "type": "array"
          },
          "contract_version": {
            "const": "secondary-tools-v2",
            "title": "Contract Version"
          },
          "evidence": {
            "$ref": "#/components/schemas/CandidateEvidenceEnvelope"
          },
          "request_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{11,63}$",
            "title": "Request Id",
            "type": "string"
          },
          "screening_kind": {
            "const": "internal_ai_document_screen",
            "title": "Screening Kind"
          },
          "summary": {
            "maxLength": 1200,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "request_id",
          "screening_kind",
          "summary",
          "claims",
          "evidence"
        ],
        "title": "StructLintResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Optional API key. Anonymous requests fall back to the free tier (60 req/min). Provide an X-API-Key header to access higher tiers.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Structural Team",
      "url": "https://structural.bytedance.city/"
    },
    "description": "跨领域结构同构搜索引擎 — Structural Isomorphism API.\n\nPublic surface for the Structural cross-domain knowledge engine. Endpoints are grouped by capability (Ask / Phase / Phenomenon / Pricing / Newsletter / Admin). Rate limits are tier-based; supply an `X-API-Key` header to promote a request beyond the anonymous free-tier bucket.",
    "license": {
      "name": "Proprietary"
    },
    "title": "Structural",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "post": {
        "operationId": "search_phenomena_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Phenomena",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/struct-lint": {
      "post": {
        "description": "Extract structural claims + failure modes from a document.\n\nResponses:\n  200 — {\"summary\", \"claims\": [...]}\n  400 — empty or over-long document\n  503 — no LLM key configured / LLM call failed",
        "operationId": "struct_lint_api_struct_lint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructLintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructLintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Struct Lint",
        "tags": [
          "struct-lint"
        ]
      }
    },
    "/api/struct-lint/stream": {
      "post": {
        "description": "SSE variant of /struct-lint — emits progress while the lint runs.\n\nEvents (mirrors api/analyze.py's naming):\n  meta     — {\"max_doc_chars\": N} sent immediately so the client knows\n             the stream is alive (fast first byte).\n  progress — {\"stage\", \"message\", ...} for each pipeline stage.\n  done     — {\"result\": {\"summary\", \"claims\": [...]}} final payload.\n  error    — {\"error\", \"message\"} terminal failure.\n\nValidation errors (empty / too-long doc, no LLM key) are surfaced as a\nterminal `error` event rather than an HTTP status, because by the time\nthe generator runs the 200 + headers are already sent.",
        "operationId": "struct_lint_stream_api_struct_lint_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructLintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Struct Lint Stream",
        "tags": [
          "struct-lint"
        ]
      }
    },
...
  },
...
}